Update slackbot-guysgals-reasoning.md (#5)
* Added 1st version of the reasoning md file, WiP
* Updated reasoning, included bit.ly link to half the responses
* Added 1st rule of DevOps
* Adjusted spacing for the headings + link
* Less pushy responses/objectionable suggestions.

I feel like a big part of the issue some have with the bot is due to it
suggesting what they mean. This asks them to consider why. I also think
it's best to explain why it's happening and link to the FAQ every time.

Finally, I've removed the suggestions people have complained about most,
legitimately or not (such as Gang etc).

* Update slackbot-guysgals-reasoning.md
* Remove duplicate responses
* Update slackbot-guysgals-reasoning.md