Channel links: Replace #others with #learn-social-issues (#8)
`#others` is probably not a good place to send people too. It's intended to be among other things a safe haven for "others" in the community, who probably don't want an influx of people who don't agree with this.

`#learn-social-issues` on the other hand, should be a safe educational space to direct people too.