Make responses clearer
When I see a bot response, it makes it seem like I'm not considering everyone, and I should be. The italics don't make it clear that I should use the italicized word instead of guys or gals.

This change makes both words look the same through both being quoted, and refactors the grammar to make it clear that the word "everyone" should be used instead.

Hope you find this helpful.